#52: Refactor DB entity
