+++ src/modules/common/db/entities/Task.entity.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-classes-per-file */
 import {
   Entity,
   PrimaryGeneratedColumn,
@@ -7,73 +6,22 @@ import {
   JoinColumn,
   CreateDateColumn,
   UpdateDateColumn,
-  DeleteDateColumn,
-  OneToMany
+  DeleteDateColumn
 } from 'typeorm'
 import { type ITaskCreateUpdateParams } from '@common/types'
-
-@Entity()
-export class DayPlan {
-  @PrimaryGeneratedColumn()
-  id: number
-
-  @Column('text')
-  day: string
-
-  @OneToMany(() => Task, (task) => task.id)
-  tasks: Task[]
-
-  @CreateDateColumn({ type: 'date' })
-  createdAt?: Date
-
-  @UpdateDateColumn({ type: 'date' })
-  updatedAt?: Date
-
-  @DeleteDateColumn({ type: 'date' })
-  deletedAt?: Date
-}
-
-@Entity()
-export class Goal {
-  @PrimaryGeneratedColumn()
-  id: number
-
-  @Column('text')
-  name: string
-
-  @Column('text')
-  description: string
-
-  @Column('text')
-  type: string
-
-  @Column({ type: 'float' })
-  progress: number
-
-  @OneToMany(() => Task, (task) => task.GoalId)
-  tasks: Task[]
-
-  @CreateDateColumn({ type: 'date' })
-  createdAt: Date
-
-  @UpdateDateColumn({ type: 'date' })
-  updatedAt: Date
-
-  @DeleteDateColumn({ type: 'date' })
-  deletedAt: Date
-}
+import {DayPlan} from "@common/db/entities/DayPlan.entity";
 
 @Entity()
 export class Task {
   constructor(params: ITaskCreateUpdateParams) {
     if (params) {
-      const { title, priority, taskType, status, startDate, endDate, description, duration, breakDuration } = params
+      const { title, priority, type, status, startDate, endDate, description, duration, breakDuration } = params
       this.status = status
       this.startDate = startDate
       this.endDate = endDate
       this.title = title
       this.priority = priority
-      this.taskType = taskType
+      this.type = type
       if (description) this.description = description
       if (duration) this.duration = duration
       if (breakDuration) this.breakDuration = breakDuration
@@ -93,7 +41,7 @@ export class Task {
   priority: string
 
   @Column('text')
-  taskType: string
+  type: string
 
   @Column('text')
   status: string
@@ -110,13 +58,9 @@ export class Task {
   @Column('date', { nullable: true })
   endDate: Date
 
-  @ManyToOne(() => Goal, (goal) => goal.tasks)
-  @JoinColumn({ name: 'GoalId' })
-  GoalId: Goal
-
   @ManyToOne(() => DayPlan, (dayPlan) => dayPlan.tasks)
   @JoinColumn({ name: 'dayPlanId' })
-  dayPlanId: DayPlan
+  dayPlan: DayPlan
 
   @CreateDateColumn({ type: 'date' })
   createdAt?: Date
