+++ src/modules/common/db/entities/index.ts
@@ -1,3 +1,4 @@
 export * from './Notification.entity'
 export * from './Task.entity'
 export * from './User.entity'
+export * from './DayPlan.entity'
